# docs/swagger.yaml
basePath: /
definitions:
...
  entity.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  request.RegisterRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  response.ErrorResponseAuth:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  response.LoginSuccessResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/entity.User'
    type: object
...
  response.RefreshSuccessResponse:
    properties:
      access_token:
        type: string
    type: object
  response.RegisterSuccessResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/entity.User'
    type: object
...
host: localhost:3100
info:
  contact: {}
  description: API for auth service
  title: Auth Service API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns user information along with a
        new access token. A new refresh token is set as an HTTP-only cookie, and any
        existing refresh token in the cookie is invalidated.
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/response.LoginSuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponseAuth'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.ErrorResponseAuth'
      summary: Log in an existing user
      tags:
      - auth
...
  /refresh:
    post:
      description: Uses a refresh token to generate a new access token and a new refresh
        token. Refresh token is set as an HTTP-only cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed tokens
          schema:
            $ref: '#/definitions/response.RefreshSuccessResponse'
        "401":
          description: Refresh token required or invalid/expired refresh token
          schema:
            $ref: '#/definitions/response.ErrorResponseAuth'
      summary: Refresh access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns user information along with
        an access token. A refresh token is set as an HTTP-only cookie.
      parameters:
      - description: User Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered
          schema:
            $ref: '#/definitions/response.RegisterSuccessResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponseAuth'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponseAuth'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"